+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - changed behaviour of VcsDataTable component: footer is now visible as long as number of items > min(itemsPerPageArray)
 - fixed a bug where a panel added to the PanelManager with `resizable:false` would appear on top of the map instead of besides it
 - changed behaviour of compass on mobile: click now aligns map to the North
+- added a new callback: `ToggleToolbarButtonCallback`
 
 # 6.1.13